requirements.txt
django==4.2.4
djangorestframework==3.14.0
Pillow==8.0.1
jinja2==3.0
gunicorn==20.1.0
contextvars
typing-extensions==4.1.1
aiohttp==3.8.1
click==8.0.4
wheel==0.41.1
multidict==4.5
djangorestframework-jwt==1.11.0
pytest==7.4.0
setuptools==68.0.0
psycopg2-binary==2.9.7
postgresql
djangorestframework_simplejwt==4.1.1
# postgresql-contrib==14.238 //should check version
